Revert "[FEAT/#151] "설정 > 알림설정" 페이지 구현"
