--- app/src/main/java/com/sopt/clody/data/remote/dto/response/ResponseNotificationInfoDto.kt
@@ -1,11 +1,0 @@
-package com.sopt.clody.data.remote.dto.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseNotificationInfoDto(
-    @SerialName("isDiaryAlarm") val isDiaryAlarm: Boolean,
-    @SerialName("isReplyAlarm") val isReplyAlarm: Boolean,
-    @SerialName("time") val time: String
-)
