+++ app/src/main/java/com/sopt/clody/data/repository/NotificationRepository.kt
@@ -1,10 +1,8 @@
 package com.sopt.clody.data.repository
 
 import com.sopt.clody.data.remote.dto.request.RequestSendNotificationDto
-import com.sopt.clody.data.remote.dto.response.ResponseNotificationInfoDto
 import com.sopt.clody.data.remote.dto.response.ResponseSendNotificationDto
 
 interface NotificationRepository {
-    suspend fun getNotificationInfo(): Result<ResponseNotificationInfoDto>
     suspend fun sendNotification(requestSendNotificationDto: RequestSendNotificationDto): Result<ResponseSendNotificationDto>
 }
