+++ app/src/main/java/com/sopt/clody/data/repositoryimpl/NotificationRepositoryImpl.kt
@@ -2,7 +2,6 @@ package com.sopt.clody.data.repositoryimpl
 
 import com.sopt.clody.data.remote.datasource.NotificationDataSource
 import com.sopt.clody.data.remote.dto.request.RequestSendNotificationDto
-import com.sopt.clody.data.remote.dto.response.ResponseNotificationInfoDto
 import com.sopt.clody.data.remote.dto.response.ResponseSendNotificationDto
 import com.sopt.clody.data.repository.NotificationRepository
 import com.sopt.clody.presentation.utils.extension.handleApiResponse
@@ -11,13 +10,6 @@ import javax.inject.Inject
 class NotificationRepositoryImpl @Inject constructor(
     private val notificationDataSource: NotificationDataSource
 ) : NotificationRepository {
-    override suspend fun getNotificationInfo(): Result<ResponseNotificationInfoDto> {
-        return runCatching {
-            val response = notificationDataSource.getNotificationInfo()
-            response.handleApiResponse().getOrThrow()
-        }
-    }
-
     override suspend fun sendNotification(requestSendNotificationDto: RequestSendNotificationDto): Result<ResponseSendNotificationDto> {
         return runCatching {
             val response = notificationDataSource.sendNotification(requestSendNotificationDto)
