+++ app/src/main/java/com/sopt/clody/presentation/ui/setting/component/AccountManagementNicknameOption.kt
@@ -1,11 +1,10 @@
-package com.sopt.clody.presentation.ui.setting.accountmanagement.component
+package com.sopt.clody.presentation.ui.setting.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -17,32 +16,33 @@ import com.sopt.clody.R
 import com.sopt.clody.ui.theme.ClodyTheme
 
 @Composable
-fun AccountManagementLogoutOption(
-    userEmail: String,
-    updateLogoutDialog: (Boolean) -> Unit
+fun AccountManagementNicknameOption(
+    userName: String,
+    updateNicknameChangeBottomSheet: (Boolean) -> Unit
 ) {
     Row(
         modifier = Modifier
-            .padding(top = 12.dp, bottom = 24.dp, start = 22.dp, end = 24.dp)
+            .padding(horizontal = 24.dp, vertical = 24.dp)
     ) {
         Image(
-            painter = painterResource(id = R.drawable.img_account_management_kakao),
-            modifier = Modifier
-                .size(24.dp),
+            painter = painterResource(id = R.drawable.ic_account_management_clover),
             contentDescription = null
         )
         Spacer(modifier = Modifier.width(10.dp))
         Text(
-            text = userEmail,
+            text = userName,
+            style = ClodyTheme.typography.body1SemiBold
+        )
+        Text(
+            text = stringResource(R.string.account_management_nickname),
             style = ClodyTheme.typography.body1Medium
         )
         Spacer(modifier = Modifier.weight(1f))
         Text(
-            text = stringResource(R.string.account_management_logout_button),
-            modifier = Modifier.clickable(onClick = { updateLogoutDialog(true) }),
+            text = stringResource(R.string.account_management_nickname_change_button),
+            modifier = Modifier.clickable(onClick = { updateNicknameChangeBottomSheet(true) }),
             color = ClodyTheme.colors.gray05,
             style = ClodyTheme.typography.body4Medium
         )
-
     }
 }
