+++ app/src/main/java/com/sopt/clody/presentation/ui/setting/component/NotificationSettingTime.kt
@@ -1,8 +1,7 @@
-package com.sopt.clody.presentation.ui.setting.notificationsetting.component
+package com.sopt.clody.presentation.ui.setting.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -10,7 +9,6 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
@@ -28,12 +26,7 @@ fun NotificationSettingTime(
         modifier = Modifier
             .fillMaxWidth()
             .wrapContentHeight()
-            .padding(horizontal = 24.dp)
-            .clickable (
-                onClick = {updateNotificationTimePicker(true) },
-                indication = null,
-                interactionSource = remember { MutableInteractionSource() }
-            ),
+            .padding(horizontal = 24.dp),
         verticalAlignment = Alignment.CenterVertically
     ) {
         Text(
@@ -49,7 +42,8 @@ fun NotificationSettingTime(
         )
         Image(
             painterResource(id = R.drawable.ic_setting_next),
-            contentDescription = "pick the alarm time"
+            contentDescription = "pick the alarm time",
+            modifier = Modifier.clickable { updateNotificationTimePicker(true) }
         )
     }
 }
