+++ app/src/main/java/com/sopt/clody/presentation/ui/setting/screen/AccountManagementViewModel.kt
@@ -1,4 +1,4 @@
-package com.sopt.clody.presentation.ui.setting.accountmanagement.screen
+package com.sopt.clody.presentation.ui.setting.screen
 
 import android.content.Context
 import android.content.Intent
